test(frontend):Fixed test issues
